finished

+++ components/Dashboard.js
@@ -1,14 +1,80 @@
+"use client";
 import { Fugaz_One } from "next/font/google";
+import React, { useEffect, useState } from "react";
 import Calendar from "./Calendar";
+import { useAuth } from "@/context/AuthContext";
+import { doc, setDoc } from "firebase/firestore";
+import { db } from "@/firebase";
+import Login from "./Login";
+import Loading from "./Loading";
 
 const fugaz = Fugaz_One({ subsets: ["latin"], weight: ["400"] });
 
 export default function Dashboard() {
+  const { currentUser, userDataObj, setUserDataObj, loading } = useAuth();
+  const [data, setData] = useState({});
+  const now = new Date();
+
+  function countValues() {
+    let totalNumOfDays = 0;
+    let sumMoods = 0;
+
+    for (let year in data) {
+      for (let month in data[year]) {
+        for (let day in data[year][month]) {
+          let daysMood = data[year][month][day];
+
+          totalNumOfDays++;
+          sumMoods += daysMood;
+        }
+      }
+    }
+    return {
+      num_days: totalNumOfDays,
+      average_mood: sumMoods / totalNumOfDays,
+    };
+  }
+
   const statuses = {
-    num_days: 14,
-    time_remaining: "13:14:26",
-    date: new Date().toDateString(),
+    ...countValues(),
+    time_remaining: `${23 - now.getHours()}H ${60 - now.getMinutes()}M`,
   };
+
+  async function handleSetMood(mood) {
+    const day = now.getDate();
+    const month = now.getMonth();
+    const year = now.getFullYear();
+
+    try {
+      const newData = { ...userDataObj };
+      if (!newData?.[year]) {
+        newData[year] = {};
+      }
+      if (!newData?.[year][month]) {
+        newData[year][month] = {};
+      }
+      newData[year][month][day] = mood;
+      // update current state
+      setData(newData);
+      // update global state
+      setUserDataObj(newData);
+      // update firebase
+      const docRef = doc(db, "users", currentUser.uid);
+      const res = await setDoc(
+        docRef,
+        {
+          [year]: {
+            [month]: {
+              [day]: mood,
+            },
+          },
+        },
+        { merge: true }
+      );
+    } catch (error) {
+      console.log("Failed to set data: ", error);
+    }
+  }
 
   const moods = {
     "&*@#$": "😭",
@@ -18,17 +84,33 @@ export default function Dashboard() {
     Elated: "😍",
   };
 
+  useEffect(() => {
+    if (!currentUser || !userDataObj) {
+      return;
+    }
+    setData(userDataObj);
+  }, [currentUser, userDataObj]);
+
+  if (loading) {
+    return <Loading />;
+  }
+
+  if (!currentUser) {
+    return <Login />;
+  }
+
   return (
     <div className="flex flex-col flex-1 gap-8 sm:gap-12 md:gap-16">
       <div className="grid grid-cols-3 bg-indigo-50 text-indigo-500 p-4 gap-4 rounded-lg">
         {Object.keys(statuses).map((status, index) => {
           return (
             <div key={index} className="flex flex-col gap-1 sm:gap-2">
-              <p className="font-medium uppercase text-xs sm:text-sm truncate">
+              <p className="font-medium capitalize text-xs sm:text-sm truncate">
                 {status.replaceAll("_", " ")}
               </p>
               <p className={"text-base sm:text-lg truncate " + fugaz.className}>
                 {statuses[status]}
+                {status === "num_days" ? "🔥" : ""}
               </p>
             </div>
           );
@@ -45,6 +127,10 @@ export default function Dashboard() {
         {Object.keys(moods).map((mood, index) => {
           return (
             <button
+              onClick={() => {
+                const currentMoodValue = index + 1;
+                handleSetMood(currentMoodValue);
+              }}
               className={
                 "p-4 px-5 rounded-2xl purpleShadow duration-200 bg-indigo-50 hover:bg-indigo-100 text-center flex flex-col items-center gap-2 flex-1"
               }
@@ -63,7 +149,7 @@ export default function Dashboard() {
           );
         })}
       </div>
-      <Calendar />
+      <Calendar completeData={data} handleSetMood={handleSetMood} />
     </div>
   );
 }
